Removing expiration read trigger
In RavenDB, when a document is expired it stops being visible immediately (for both queries and load operations) and deleted on a first chance basis (depending on indexing performance).

For various technical reasons (commit 00c8201 and the fact we are not holding the criteria for expiration in the metadata), it is going to cost us a bit to do the same type of blocking.

That's sort of an edge case really, but worth noting. Side effect is there will be a twilight zone around the expiration threshold (which is constantly moving) where messages may or may not be visible, and conversations will/might appear partial (which will be the case anyway, actually).