On critical error shutdown servicecontrol
Unfortunately to do a clean shutdown (not using Environment.FailFast) there seems to be a race condition, so i had to ensure that a call to RaiseCriticalError is done from a different thread then the existing one that the de-queuer is on.
I wonder if this issue also exists in the core!