Add a DB maintenance mode to ServiceControl
Added the-maintenance  command line switch to allow ServiceControl to be
started with reduced functionality.  In this mode it serves up the
RavenDB Web UI but disables the Raven Bundles and the periodic deletion
code and does not start the bus so message processing and the REST API
are also disabled.