Upgrade to Raven v3
This includes moving the database under Databases\ServiceControl, so before the db was in eg C:\ProgramData\Particular\ServiceControl\localhost-33333, now the actual db is located in C:\ProgramData\Particular\ServiceControl\localhost-33333\Databases\ServiceControl.

We also taking advantage of v3 supporting OWIN and this means that we use a single HttpListeners to serve both ServiceControl and Raven studio.