Both instances may be publishing this event
In a message-driven pubsub scenario, it doesn't matter because we only subscribe to one source. In a brokered transport, we run into race conditions as to which two custom check results arrive first, and even whether the next two arrive between stating that the test is done and the endpoint shutting down.