Fix ServiceControl is stopped detection logic
When stopping the ServiceControl service, the Windows Service itself
stops fairly fast--usually much less than 1 minute. However, the process
doesn't always teardown in the currently allotted 5 seconds.

However, even if the wait time were longer, the code to detect whether
or not the process has exited still may not work as intended. The Task
which is started spins a while loop waiting for the process to exit. If
the process is no longer detected, the while loop terminates, as does
the task. This would result in the call to Task.Wait(n) to return
'true'. However, in the case where the process does not exit in the
allotted 5 seconds, Task.Wait(n) would return 'false' due to the task
not completing successfully within the given timeout.

It is this return value which should signal whether or not the upgrade
process can proceed. But the return value from Task.Wait(n) is never
captured. Instead, 'true' is simply returned, which always results in
ServiceControl thinking it's OK to proceed with an upgrade.

Instead of returning 'true', return the return value from Task.Wait(n).