Don't attempt upgrade if service is not stopped
Even after the fix for detecting whether or not the ServiceControl
service is stopped (and, apparently, whether or not the process has
exited), all of the installers, upon upgrade when the check to see if
the service is stopped is performed, if the service is not stopped, then
an error is logged. BUT, the upgrade process continues. If the
ServiceControl service (and underlying process) have not stopped, this
obviously will result in an upgrade failure (usually due to a locked
Autofac.dll file).

Instead, if the service is detected as not stopped, log the error, but
also return false from the Upgrade method.