Use Microsoft.Data.SqlClient for Validation
The Sql Transport uses Microsoft.Data.SqlClient for interacting with Sql
Server databases. This is good and necessary to be able to connect to
Azure Sql using Azure Active Directory credentials. This works fine
having installed a Service Control instance with the Sql Transport.

Installing an instance, though, fails via both the Service Control
Management tool and the PowerShell modules. This is because the
ServiceControlInstallerEngine library validates connection strings by
attempting to connect to the database specified. It does so using
System.Data.SqlClient, which does not support the use of Azure Active
Directory authentication types in the connection string.

To make a seamless experience when using Sql for transport, it's
necessary for such connection strings to successfully validate.